% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdev24-package.R
\docType{package}
\name{pkgdev24-package}
\alias{pkgdev24}
\alias{pkgdev24-package}
\title{pkgdev24: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeffreyad.github.io/pkgdev24/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
